Agg builder accessibility fixes
- Getters for DateHisto `interval` and `order` should return a
long, not double
- Add getter for the filter in a FilterAgg
- Add getters for subaggs / pipelines in base AggregationBuilder

+++ core/src/main/java/org/elasticsearch/search/aggregations/bucket/histogram/DateHistogramAggregationBuilder.java
@@ -164,7 +164,7 @@ public class DateHistogramAggregationBuilder
     }
 
     /** Get the current interval in milliseconds that is set on this builder. */
-    public double interval() {
+    public long interval() {
         return interval;
     }
 
@@ -196,7 +196,7 @@ public class DateHistogramAggregationBuilder
     }
 
     /** Get the offset to use when rounding, which is a number of milliseconds. */
-    public double offset() {
+    public long offset() {
         return offset;
     }
